Make the mixins for ExtentViewDownsize work
