+++ src/main/java/org/spongepowered/common/mixin/core/world/MixinChunk.java
@@ -51,6 +51,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.common.util.SpongeHooks;
 import org.spongepowered.common.util.VecHelper;
+import org.spongepowered.common.world.extent.ExtentViewDownsize;
 import org.spongepowered.common.world.storage.SpongeChunkLayout;
 
 @NonnullByDefault
@@ -222,7 +223,7 @@ public abstract class MixinChunk implements Chunk {
     public Extent getExtentView(Vector3i newMin, Vector3i newMax) {
         checkBlockBounds(newMin.getX(), newMin.getY(), newMin.getZ());
         checkBlockBounds(newMax.getX(), newMax.getY(), newMax.getZ());
-        return null;
+        return ExtentViewDownsize.newInstance(this, newMin, newMax);
     }
 
     @Override
